Added zipped project directory
So the main components of this are:
app:  This contains the css, js, views, and index(login).html and the resources folder which is just where we can put imgs or bower shit for angular extensions
node_modules:  this is all the fancy node stuff that lets you inject angular into the static html.  I usually just do npm install to put all the packages in a project because I don't know enough to cherry pick the ones I need.  
tests:  We can run tests from here but I didn't write any lol.

I ran this out of intellij, but any IDE with js capabilities should be fine