"Removed" User/Group Walls
As was pointed out in one of the issues, we don't really need a group/user profile AND wall, because they're doing almost exactly the same thing right now. All that we did in this update was comment out the links to their pages (so they won't appear), making them inaccessible without a direct url. We're prolly gonna just delete the pages entirely later.