Search: enable loading a specific user/group Profile
Makes use of UserService and GroupService in app.js
to pass the name of the user/group that whose page we
want to view.

Modified UserService to return the name of a user, just
for testing, though eventually we'll want to pass the
user's id, which is what the back-end needs.

Modified userProfCtrl, groupProfCtrl so they get the
name of user/group passed via UserService/GroupService.